POWR-757 Improve change group action

+++ web/modules/custom/portland/src/Plugin/Action/ChangeGroupAction.php
@@ -8,6 +8,7 @@ use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\Core\Plugin\PluginFormInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\group\Entity\Group;
+use Drupal\group\Entity\GroupContent;
 
 /**
  * Some description.
@@ -16,7 +17,7 @@ use Drupal\group\Entity\Group;
  *   id = "change_group_action",
  *   label = @Translation("Change parent group of Content (custom action)"),
  *   type = "",
- *   confirm = TRUE,
+ *   confirm = FALSE,
  * )
  */
 class ChangeGroupAction extends ViewsBulkOperationsActionBase implements PluginFormInterface {
@@ -27,31 +28,56 @@ class ChangeGroupAction extends ViewsBulkOperationsActionBase implements PluginF
    * {@inheritdoc}
    */
   public function execute($entity = NULL) {
-    // Remove from all groups.
-    $group_contents = \Drupal::entityTypeManager()
-        ->getStorage('group_content')
-        ->loadByProperties([ 
-            'entity_id' => $entity->id(),
-            // 'gid' => $old_group_id, // The item can only belong to one group. Don't need group id.
-    ]);
-    foreach ($group_contents as $item) {
-        $item->delete();
+    /*
+    The user intention is 
+    1. Change the group the content belongs to. 
+    2. Update the group field value to be the same as 1.
+
+    if old != new
+      if new != NULL && old != NULL
+        remove from old group and add to new
+      else new != NULL && old == NULL
+        add to new group
+      else new == NULL && old != NULL
+        remove from old group
+    */
+    $new_group_id = $this->configuration['new_group']; //string
+    $new_group = ($new_group_id == 0) ? NULL : Group::load($new_group_id);
+    $new_group_type = ($new_group == NULL) ? NULL : $new_group->getGroupType()->id();
+    $old_group_id_array = self::getGroupIdsByEntity($entity);
+    $old_group_id = (count($old_group_id_array) > 0) ? $old_group_id_array[0] : 0;
+
+    // Services can only be added to Bureaus and Programs.
+    if( $entity->bundle() == 'city_service' && 
+      ($new_group_type != 'bureau_office' && $new_group_type != 'program') ) 
+    {
+      // Make sure Group field is the same as old group
+      if( $entity->hasField('field_group') ) $entity->field_group->entity = $old_group;
+      $entity->save();
+      return $this->t('Service can only be in Bureaus/Offices or Programs.');
     }
 
-    if($this->configuration['new_group'] == 0) {
-      // Update the Group field value
-      $entity->field_group = NULL;
-      $entity->save();
+    if($old_group_id != $new_group_id) {
+      if($old_group_id != 0) {
+        // Remove from all groups.
+        $group_contents = \Drupal::entityTypeManager()
+            ->getStorage('group_content')
+            ->loadByProperties([ 
+                'entity_id' => $entity->id(),
+                // 'gid' => $old_group_id, // The item can only belong to one group. Don't need group id.
+        ]);
+        foreach ($group_contents as $item) {
+            $item->delete();
+        }
+      }
+      if($new_group_id != 0) {
+        // Add to new group
+        $new_group->addContent($entity, 'group_'.$entity->getEntityTypeId().':'.$entity->bundle());
+      }
     }
-    else {
-      // Add to new group
-      $group = Group::load($this->configuration['new_group']);
-      $group->addContent($entity, 'group_'.$entity->getEntityTypeId().':'.$entity->bundle());
 
-      // Update the Group field value
-      $entity->field_group->entity = $group;
-      $entity->save();
-    }
+    if( $entity->hasField('field_group') ) $entity->field_group->entity = $new_group;
+    $entity->save();
 
     // Don't return anything for a default completion message, otherwise return translatable markup.
     return $this->t('Group changed');
@@ -117,5 +143,19 @@ class ChangeGroupAction extends ViewsBulkOperationsActionBase implements PluginF
    */
   public function submitConfigurationForm(array &$form, FormStateInterface $form_state) {
     $this->configuration['new_group'] = $form_state->getValue('new_group');
+  }
+
+  // Get the parent group IDs of an entity
+  // https://drupal.stackexchange.com/questions/238755/how-to-get-group-ids-by-ids-of-elements-of-group-content
+  public static function getGroupIdsByEntity($entity) {
+    if($entity == NULL) return [];
+    $group_ids = [];
+
+    $group_contents = GroupContent::loadByEntity($entity);
+    foreach ($group_contents as $group_content) {
+        $group_ids[] = $group_content->getGroup()->id();
+    }
+
+    return $group_ids;
   }
 }
